Add pytest and HACS/hassfest CI, README badges and MIT license
Cover the integration with a pytest suite built on
pytest-homeassistant-custom-component: the HTTP client (status parsing,
auth/connection error mapping, setspeed query building), the config flow
(user, reconfigure, reauth), the setup and coordinator
(transient-failure tolerance, reauth on 401), the fan (speed aliases,
boost semantics, optimistic state) and the number and sensor entities
(timer countdown, power, daily energy integration and restore). 119
tests, 100% statement coverage.

Add two workflows running on pushes to main, on pull requests and on
demand: `tests.yml` (ruff lint and format checks, then pytest on Python
3.13 with a 95% coverage floor) and `validate.yml` (the HACS integration
validation and hassfest). The HACS job ignores the `brands` check, which
only applies to integrations shipped in home-assistant/brands.

Add the MIT license, required by the HACS validation, and declare the
minimum supported Home Assistant version in hacs.json.

Show the HACS, latest release, downloads, CI status and license badges
in the README, along with the two My Home Assistant redirect buttons
(open the repository in HACS, start the integration setup) and a section
on running the test suite locally.

Also apply the mechanical lint fixes ruff produced: import ordering,
formatting, a redundant int() cast around math.ceil and an obsolete
noqa. No behaviour changes.

+++ custom_components/nrf905_vmc/sensor.py
@@ -9,8 +9,8 @@ truth for the sensor.
 
 from __future__ import annotations
 
-import math
 from datetime import date, datetime, timedelta
+import math
 from typing import Any
 
 from homeassistant.components.sensor import (
@@ -125,7 +125,7 @@ class Nrf905TimerRemaining(CoordinatorEntity[Nrf905Coordinator], SensorEntity):
         if self._end is None:
             return None
         remaining = (self._end - dt_util.utcnow()).total_seconds()
-        return max(0, int(math.ceil(remaining / 60)))
+        return max(0, math.ceil(remaining / 60))
 
     @property
     def extra_state_attributes(self) -> dict[str, Any]:
